refactor: fold output adapter into builder
